# swagger.yaml
swagger: "2.0"

info:
  title: My REST API
  description: A simple REST API example
  version: 1.0.0
  contact:
    name: Oeyshik Das
    email: oeyshikdas1853@gmail.com


paths:
  /api/servers:
    get:
      summary: Retrieve all servers
      responses:
        200:
          description: OK
        404:
          description: Not Found
    post:
      summary: Create a new server
      parameters:
        - name: server
          in: body
          description: Server object to be created
          required: true
          schema:
            $ref: '#/definitions/Server'
      responses:
        201:
          description: Created
        400:
          description: Bad Request
  /api/servers/{id}:
    get:
      summary: Retrieve a server by ID
      parameters:
        - name: id
          in: path
          description: ID of the server to retrieve
          required: true
          type: string
      responses:
        200:
          description: OK
        404:
          description: Not Found
    put:
      summary: Update a server by ID
      parameters:
        - name: id
          in: path
          description: ID of the server to update
          required: true
          type: string
        - name: server
          in: body
          description: Updated server object
          required: true
          schema:
            $ref: '#/definitions/Server'
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
    delete:
      summary: Delete a server by ID
      parameters:
        - name: id
          in: path
          description: ID of the server to delete
          required: true
          type: string
      responses:
        204:
          description: No Content
        404:
          description: Not Found
  /api/servers/findByName:
    get:
      summary: Find servers by name
      parameters:
        - name: name
          in: query
          description: Name to search for
          required: true
          type: string
      responses:
        200:
          description: OK
        404:
          description: Not Found
          
definitions:
  Server:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      language:
        type: string
      framework:
        type: string